Refactored fragments into separate folder.
