+++ app/src/main/java/com/example/foodreminderapp/fragments/FoodItemListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.foodreminderapp
+package com.example.foodreminderapp.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,6 +7,10 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
+import com.example.foodreminderapp.FoodItemListAdapter
+import com.example.foodreminderapp.FoodItemListApplication
+import com.example.foodreminderapp.FoodItemListViewModel
+import com.example.foodreminderapp.FoodItemViewModelFactory
 import com.example.foodreminderapp.databinding.FragmentItemListBinding
 
 /**
