+++ app/src/main/res/layout/fragment_create_edit_item.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".CreateEditItemFragment"
+    tools:context=".fragments.CreateEditItemFragment"
     android:layout_margin="15dp">
