+++ app/src/main/res/layout/fragment_item_details.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ItemDetailFragment"
+    tools:context=".fragments.ItemDetailFragment"
     android:layout_margin="15dp">
 
     <com.google.android.material.button.MaterialButton
