+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/foodItemListFragment"
-        android:name="com.example.foodreminderapp.FoodItemListFragment"
+        android:name="com.example.foodreminderapp.fragments.FoodItemListFragment"
         android:label="@string/mainScreenTitle"
         tools:layout="@layout/fragment_item_list">
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/createEditFragment"
-        android:name="com.example.foodreminderapp.CreateEditItemFragment"
+        android:name="com.example.foodreminderapp.fragments.CreateEditItemFragment"
         android:label="@string/createEditTitle"
         tools:layout="@layout/fragment_create_edit_item">
         <argument
@@ -32,7 +32,7 @@
     </fragment>
     <fragment
         android:id="@+id/itemDetailFragment"
-        android:name="com.example.foodreminderapp.ItemDetailFragment"
+        android:name="com.example.foodreminderapp.fragments.ItemDetailFragment"
         android:label="@string/itemDetailsTitle"
         tools:layout="@layout/fragment_item_details" >
